Add raster lookup in geotiff files
This commit introduce the raster lookup file. This will take in any list
of geotiff files, and it's possible to perform elevation searches based on
geocoordinates. This will greatly simplify heightmap generation, in which
geocoordinates can be provided in order to get the elevation at that
point.

Also added .gitignore rules for python generated files.

Dependency added for GDAL