Change nodata-value handling
Previously, nodata values was not really handled, and was simply returned.
Now, nodata is by default handled as None values

Previously, nodata values could be returned from the first heightmap it
found. With this commit, the next dataset file will be attempted if a
nodata value is found. If nodata or bounds is returned from all, it will
simply return the specified nodata value which is defaulted to None