Add heightmap and heightmap maker
The heightmap class is able to create simple image-like heightmaps from
raster files. This is done by iterating over all the geo-coordinates that
is represented by each pixel. This is the reason for providing
geo-transformation for the heightmap.

Example code is provided through a runnable python file that can create a
heightmap by specifying heightmap width, height and the geo-locations of
the top-left and top right corner of the image together with the geotiff
files.

numpy dependency added

matplotlib.pyplot dependencies are added, but is only required if the plot
argument is provided to the make_heightmap.py program.