Add mesh generation and export tool
This commit introduces the mesh module and heightmap_mesh that is able to
create meshes from heightmap data, such as numpy arrays.

It also includes the obj_mesh.py module that is able to export a mesh to
.obj format. Currently only vertices, object name and faces are supported
(leaving normals and uv-mapping for a later release)

The make_mesh_from_heightmap.py tool is added as a runnable program to
take input arguments and make a mesh based on that. Currently, a numpy
heightmap must be provided as input.